idflow/tasks/gpt_researcher/gpt_researcher.py
# TYPE: GPT_RESEARCHER

from conductor.client.worker.worker_task import worker_task
import requests
import json
from typing import Dict, Any


@worker_task(task_definition_name='gpt_researcher')
def gpt_researcher(
    query: str,
    docId: str,
    max_results: int = 5,
    research_depth: str = "basic"
) -> Dict[str, Any]:
    """
    GPT Researcher task for conducting research on a given query.

    Args:
        query: The research query/topic
        docId: Document ID for tracking
        max_results: Maximum number of research results to return
        research_depth: Research depth level (basic, intermediate, advanced)

    Returns:
        Dictionary containing research results
    """
    try:
        # In a real implementation, this would integrate with GPT Researcher
        # For now, we'll simulate the research process

        # Simulate research based on query
        research_results = {
            "query": query,
            "docId": docId,
            "method": "gpt_researcher",
            "depth": research_depth,
            "results": [
                {
                    "title": f"Understanding {query.split()[0]} Technologies",
                    "summary": f"Comprehensive analysis of {query} including current trends and applications",
                    "source": "Academic Research Database",
                    "relevance_score": 0.95,
                    "key_points": [
                        f"Advanced applications of {query.split()[0]} in modern systems",
                        "Performance comparisons and benchmarks",
                        "Future development trends and predictions"
                    ]
                },
                {
                    "title": f"Industry Applications of {query.split()[-1]}",
                    "summary": f"Real-world implementations and case studies for {query}",
                    "source": "Industry Reports",
                    "relevance_score": 0.88,
                    "key_points": [
                        "Enterprise adoption patterns",
                        "Cost-benefit analysis",
                        "Implementation challenges and solutions"
                    ]
                }
            ],
            "summary": f"Research completed on '{query}' using GPT Researcher methodology",
            "confidence_score": 0.92,
            "timestamp": "2024-01-15T10:30:00Z"
        }

        return research_results

    except Exception as e:
        return {
            "error": f"GPT Researcher failed: {str(e)}",
            "query": query,
            "docId": docId,
            "method": "gpt_researcher"
        }
