pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "idflow"
version = "0.1.0"
description = "A document management and workflow system"
requires-python = ">=3.8"
dependencies = [
    "typer>=0.12",
    "pydantic>=2",
    "PyYAML>=6",
    "conductor-python",
    "litellm>=1.0.0",
    "keybert>=0.9.0"
]

[project.scripts]
idflow = "idflow.__main__:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["idflow*"]

[tool.pytest.ini_options]
testpaths = ["tests", "tasks"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests"
]
